added src/Option.js
