+++ src/Dropdown.js
@@ -1,30 +1,9 @@
 import React, { Component } from 'react';
 import ReactDOM from 'react-dom';
 import classNames from 'classnames';
-import { getOptionLabel, getOptionValue } from './helpers';
+import Option from './Option';
 
 const DEFAULT_PLACEHOLDER_STRING = 'Select...';
-
-export const Option = props => {
-  const { option, selected, onSelect, className } = props;
-  const value = getOptionValue(option);
-  const label = getOptionLabel(option);
-  const isSelected = value === selected;
-  let optionClassName = `Dropdown-option ${className || ''}`;
-
-  optionClassName = `${optionClassName} ${isSelected ? 'is-selected' : ''}`;
-
-  return (
-    <div
-      className={optionClassName.trim()}
-      onMouseDown={e => onSelect(e, value, label)}
-      onClick={e => onSelect(e, value, label)}
-      role="option"
-      aria-selected={String(isSelected)}>
-      {label}
-    </div>
-  );
-};
 
 class Dropdown extends Component {
   constructor (props) {
